+++ packages/shared/widget-plugin-grid/jest.config.cjs
@@ -5,6 +5,10 @@ module.exports = {
             "@swc/jest",
             {
                 jsc: {
+                    parser: {
+                        syntax: "typescript",
+                        tsx: true
+                    },
                     transform: {
                         react: {
                             runtime: "automatic"
@@ -15,6 +19,7 @@ module.exports = {
         ]
     },
     moduleNameMapper: {
+        "^big\\.js$": "<rootDir>/../../../node_modules/big.js/big.js",
         "(.+)\\.js": "$1"
     },
     extensionsToTreatAsEsm: [".ts"]