+++ packages/shared/widget-plugin-grid/src/pagination/PageControl.service.ts
@@ -7,7 +7,10 @@ import { SetPageAction, SetPageSizeAction } from "./pagination.model";
 export class PageControlService implements GridPageControl {
     constructor(
         private gate: DerivedPropsGate<{
+            dynamicPage?: EditableValue<Big>;
+            dynamicPageSize?: EditableValue<Big>;
             totalCountValue?: EditableValue<Big>;
+            loadedRowsValue?: EditableValue<Big>;
         }>,
         private setPageSizeAction: SetPageSizeAction,
         private setPageAction: SetPageAction
@@ -15,6 +18,9 @@ export class PageControlService implements GridPageControl {
 
     setPageSize(pageSize: number): void {
         this.setPageSizeAction(pageSize);
+        const value = this.gate.props.dynamicPageSize;
+        if (!value || value.readOnly) return;
+        value.setValue(new Big(pageSize));
     }
 
     setPage(page: number): void {
@@ -23,6 +29,12 @@ export class PageControlService implements GridPageControl {
 
     setTotalCount(count: number): void {
         const value = this.gate.props.totalCountValue;
+        if (!value || value.readOnly) return;
+        value.setValue(new Big(count));
+    }
+
+    setLoadedRows(count: number): void {
+        const value = this.gate.props.loadedRowsValue;
         if (!value || value.readOnly) return;
         value.setValue(new Big(count));
     }